Class Work Clean up
Added documentation to the class work.  Reworked the multivariate part
of the function so that one can choose to have preferences based on 1,2
or 3 distributions simply by the number of mean vectors and vcov
matrices provided.  Also made the user specified normals take variances
instead of standard deviations (this is what the worksheet says to do).