Code unification
Edited the code so that Dalston and Myunghoon’s section work with one
another.  Made the following changes: 1. changed “affil” to
“affiliation” in the VoterAffiliate function.  This makes the code more
consistent, and easier to understand.  2. changed the values in affil
from “party1” and “party2” simply to “1” and “2”.  This makes the
Voters object continue to be a numeric matrix, instead of a character
matrix, which I believe will be useful downstream as we further
manipulate this object.  3. added an if statement at the end of the
VoterAffiliate function.  I noticed that if one ran the VoterAffilliate
function on the same voters object it would simply keep adding columns.
 Now, the Voter affiliate function either adds a third column, or if a
third column already exists, it simply replaces this column with the
new affiliation object.  This way we won’t accidentally create Voters
objects with 2 or more “affiliation” columns with the same name.  4.
Changed the ifelse() functions in the visual function to look for 1
instead of “party1” (in line with change 2 mentioned above). 5. Changed
the PartyStarter function to name the parties “party1” and “party2”, so
that it is in accord with Myunghoon’s code. 6. changed the
PartyRelocator function to run on apply loops that subset the data
according to the structure of the Voters objects produced by
VoterAffiliate (the old method was dependent on variable names and
Voters being a data frame, the new method simply needs the columns to
be in the right order).  7. I edited the PartyRelocator documentation
to account for these changes.