ElectoralSimulations function update
This commit contains code and documentation for the full electoral
simulation function.  The function as written does not print any plots,
and only returns output from the final simulated election.  This update
also required a minor change to the VoterAffiliate function - I changed
the input from party to Parties.  This was breaking the
ElectoralSimulations function (I’m not exactly sure why though…)