Electoral Simulations Update
This commit contains documentation for the visualization part of the
electoral simulation function.  It also includes an update to the
ElectoralSimulations function which makes it plot everything correctly
when nsims=1 and also to break if nsims is not 1 or greater.

This update also changes the argument names for the VoterDistribution
and Party Starter functions.  All of their arguments now have a V or a
P in front of them, so that the user can specify the parameters for the
parties or voters or both, as desired in the Electoral Simulations
function.