Color Plotting and other minor changes
This commit changes the following:

Adds several more colors to our “color palette”. Also removes the
palette command from the ElectoralSimulations function.  The reason for
this is that the palette command operates on the entire Rsession, not
just the environment in which it is located.  Weird, right?

Changes the pch for the party starting point from 20 to 19.  This makes
it a larger circle, and thus, easier to differentiate from other points.

added the argument “drop=FALSE” to the apply function in the Party
relocator.  This stops the function from breaking when there is only a
single voter supporting a party.

Reworked the Electoral simulations function to overlay voter points on
top of one other.  This allows us to somewhat visualize which voters
are switching parties in every iteration.  The key line of code for
making this work is line 272.

Changed the cex for party points that aren’t the first or last election
to 1 from 1.2.  Again, this allows for easier differentiation between
first, intermediate, and last, elections.

Condensed two if loops into a single if loop with another if loop
nested inside of it at line 312

throughout I added or edited several comments.