Expand Grid and documentation
Finalized the expand grid section.  Include a model describing what is
going on with the expand grid and code that makes a 3d plot to
investigate the comparative statics.

Also added documentation to each section of the code saying exactly
what part of the homework each function/section of code addresses.